Removed logical inconsistency.
When several physical quantities are decleared with the same unit the
reasoner will infer them to be equivalent due to the inverse
functional property of the hasUnit relation.  To avoid this, only one
out of a set of physical quantities with same dimensionality should be defined with a unit.  There rest of the quantities should subclass it.

Fixing this required added the quantities InverseLength and
EnergyPerTemperature.

Grouped dimensionnally equivalent quantities i