Updated labels of sub-properties of rdfs:seeAlso
'Match' is used for resolvable URLs to corresponding entity in another
ontology.

'Entry' is used for resolvable URLs to a human readable resource describing
the subject.

'Ref' is used for non-resolvable reference to a human readable resource
describing the subject.

Also made rdfs:isDefinedBy a sub-property of rdfs:seeAlso, which is how it
is defined in the RDF Schema.