Updated literals ending with a double quote.
At least owlapi escapes the end quote, but not the start quote, in
literals ending with a double quote when serialising to turtle. That
makes the quotes in the literal not matching, confusing editors and
other tools.  The solution is to end literals with something else than
a double quote or escape them explicitely.