Reintroduced disjoint unions (where they could be introduces).
Cleaned also up the mess in the description of bosons and make a clear separation between the following properties:
   - elementary/non-elementary
   - whether the boson is a carrier of fundamental forces (guage boson)
   - spin (scalar/vector/tensor bosons)
   - mass (massless/massive)
   - parity (pseudovector boson)

Note that Wikipedia messes this slightly up, which complicated the work.

Introduced ElementaryBoson, TensorBoson and PseudoVectorBoson to make the taxonomy cleaner.

By cleaning this up removed the equivalence between Higgs boson and scalar boson as well as between graviton and tensor boson.