Corrected changes to the EMMO logo - ensure that the middle circle is white (#354)
Corrected the change to the EMMO logo in commit ed0d63303c7a, where the
middle circle was made transparent.

This PR changes the middle circle back to be solid white.

Georg came up with many reasons for the 4 circles when we created the
logo
- EMMO is 4D 
- EMMO has 4 signs (there Exists a Whole World Ontology)
- semiotics: referent-interpreter-sign and semiosis in the centre
- Combining the colours of the three outer circles RGB gives white in
the centre

For more points, see
https://github.com/emmo-repo/EMMO/blob/master/doc/about-name-logo.md

---------

Co-authored-by: hothello <roscioni@gmail.com>