Made the IRI of hasORCID human readable.
Motivation: Users may want to look at the contributors.ttl file as-is.
Replacing the numerical IRI for hasORCID to emmo:hasORCID makes the file much
easier to read.

Furthermore, the contributors.ttl file defines only the contributing people
and organisations, but is not part of EMMO per see. There is therefor no
strong argument to keep the numerical IRI for hasORCID.